refactored the package name to connectors.integration
