+++ postgres-connector/src/main/java/org/odpi/openmetadata/adapters/connectors/integration/postgres/PostgresDatabaseConnector.java
@@ -1,4 +1,4 @@
-package org.odpi.openmetadata.adapters.connectors.datastore.postgres;
+package org.odpi.openmetadata.adapters.connectors.integration.postgres;
 
 /* SPDX-License-Identifier: Apache-2.0 */
 /* Copyright Contributors to the ODPi Egeria project. */
@@ -6,9 +6,13 @@ package org.odpi.openmetadata.adapters.connectors.datastore.postgres;
 import org.odpi.openmetadata.accessservices.datamanager.metadataelements.DatabaseColumnElement;
 import org.odpi.openmetadata.accessservices.datamanager.metadataelements.DatabaseElement;
 import org.odpi.openmetadata.accessservices.datamanager.properties.*;
-import org.odpi.openmetadata.adapters.connectors.datastore.postgres.ffdc.PostgresConnectorAuditCode;
-import org.odpi.openmetadata.adapters.connectors.datastore.postgres.ffdc.PostgresConnectorErrorCode;
-import org.odpi.openmetadata.adapters.connectors.datastore.postgres.properties.*;
+import org.odpi.openmetadata.adapters.connectors.integration.postgres.ffdc.PostgresConnectorAuditCode;
+import org.odpi.openmetadata.adapters.connectors.integration.postgres.ffdc.PostgresConnectorErrorCode;
+import org.odpi.openmetadata.adapters.connectors.integration.postgres.properties.PostgresColumn;
+import org.odpi.openmetadata.adapters.connectors.integration.postgres.properties.PostgresDatabase;
+import org.odpi.openmetadata.adapters.connectors.integration.postgres.properties.PostgresForeginKeyLinks;
+import org.odpi.openmetadata.adapters.connectors.integration.postgres.properties.PostgresSchema;
+import org.odpi.openmetadata.adapters.connectors.integration.postgres.properties.PostgresTable;
 import org.odpi.openmetadata.frameworks.connectors.ffdc.ConnectorCheckedException;
 import org.odpi.openmetadata.frameworks.connectors.ffdc.InvalidParameterException;
 import org.odpi.openmetadata.frameworks.connectors.ffdc.PropertyServerException;
@@ -29,7 +33,7 @@ public class PostgresDatabaseConnector extends DatabaseIntegratorConnector {
         PostgresSourceDatabase sourceDatabase = new PostgresSourceDatabase(connectionProperties);
 
         try {
-            List < PostgresDatabase > dbs = sourceDatabase.getDabaseNames ( );
+            List <PostgresDatabase> dbs = sourceDatabase.getDabaseNames ( );
 
             for ( PostgresDatabase db : dbs ) {
 
@@ -175,7 +179,7 @@ public class PostgresDatabaseConnector extends DatabaseIntegratorConnector {
         try {
 
             PostgresSourceDatabase sourceDB = new PostgresSourceDatabase ( this.connectionProperties );
-            List < PostgresSchema > schemas = sourceDB.getDatabaseSchema( db.getName ( ) );
+            List <PostgresSchema> schemas = sourceDB.getDatabaseSchema( db.getName ( ) );
 
             for ( PostgresSchema schema : schemas ) {
                 DatabaseSchemaProperties schemaProps = new DatabaseSchemaProperties ( );
@@ -250,7 +254,7 @@ public class PostgresDatabaseConnector extends DatabaseIntegratorConnector {
      */
     private void addTablesForSchema ( PostgresSourceDatabase sourceDB , PostgresSchema schema , String schemaGUID ) throws ConnectorCheckedException {
         String methodName = "addTablesForSchema";
-        List < PostgresTable > tables;
+        List <PostgresTable> tables;
 
         try {
             /* add the schema tables */
@@ -339,7 +343,7 @@ public class PostgresDatabaseConnector extends DatabaseIntegratorConnector {
             List < PostgresTable > tables = sourceDB.getTables ( schema );
 
             for ( PostgresTable t : tables ) {
-                List < PostgresForeginKeyLinks > foreginKeys = sourceDB.getForeginKeyLinksForTable ( t.getTable_name ( ) );
+                List <PostgresForeginKeyLinks> foreginKeys = sourceDB.getForeginKeyLinksForTable ( t.getTable_name ( ) );
                 List < String > importedGuids = new ArrayList <> ( );
                 List < String > exportedGuids = new ArrayList <> ( );
 
@@ -522,7 +526,7 @@ public class PostgresDatabaseConnector extends DatabaseIntegratorConnector {
 
         try {
             List < String > primaryKeys = sourceDB.getPrimaryKeyColumnNamesForTable ( table.getTable_name ( ) );
-            List < PostgresColumn > cols = sourceDB.getColumnAttributes ( table.getTable_name ( ) );
+            List <PostgresColumn> cols = sourceDB.getColumnAttributes ( table.getTable_name ( ) );
 
             for ( PostgresColumn col : cols ) {
                 DatabaseColumnProperties colProps = new DatabaseColumnProperties ( );
