+++ postgres-connector/src/main/java/org/odpi/openmetadata/adapters/connectors/integration/postgres/PostgresDatabaseProvider.java
@@ -1,6 +1,6 @@
 /* SPDX-License-Identifier: Apache-2.0 */
 /* Copyright Contributors to the ODPi Egeria project. */
-package org.odpi.openmetadata.adapters.connectors.datastore.postgres;
+package org.odpi.openmetadata.adapters.connectors.integration.postgres;
 
 import org.odpi.openmetadata.frameworks.connectors.ConnectorProviderBase;
 import org.odpi.openmetadata.frameworks.connectors.properties.beans.ConnectorType;
