+++ postgres-connector/src/main/java/org/odpi/openmetadata/adapters/connectors/integration/postgres/ffdc/PostgresConnectorAuditCode.java
@@ -1,6 +1,6 @@
 /* SPDX-License-Identifier: Apache-2.0 */
 /* Copyright Contributors to the ODPi Egeria project. */
-package org.odpi.openmetadata.adapters.connectors.datastore.postgres.ffdc;
+package org.odpi.openmetadata.adapters.connectors.integration.postgres.ffdc;
 import org.odpi.openmetadata.frameworks.auditlog.messagesets.AuditLogMessageDefinition;
 import org.odpi.openmetadata.frameworks.auditlog.messagesets.AuditLogMessageSet;
 import org.odpi.openmetadata.repositoryservices.auditlog.OMRSAuditLogRecordSeverity;
