+++ postgres-connector/src/main/java/org/odpi/openmetadata/adapters/connectors/integration/postgres/properties/PostgresForeginKeyLinks.java
@@ -1,7 +1,7 @@
 /* SPDX-License-Identifier: Apache 2.0 */
 /* Copyright Contributors to the ODPi Egeria project. */
 
-package org.odpi.openmetadata.adapters.connectors.datastore.postgres.properties;
+package org.odpi.openmetadata.adapters.connectors.integration.postgres.properties;
 
 public class PostgresForeginKeyLinks {
